correcting norm + adding hit or miss messages

+++ main.c
@@ -14,9 +14,13 @@ int counter = 0;
 void receive_signal(int sig, siginfo_t *siginfo, void *context)
 {
     if (sig == SIGUSR2) {
-        my_putstr("enemy connected\n\n");
-        counter = -1 * (siginfo->si_pid);
-        kill(siginfo->si_pid, SIGUSR1);
+        if (counter == 0) {
+            my_putstr("enemy connected\n\n");
+            counter = -1 * (siginfo->si_pid);
+            kill(siginfo->si_pid, SIGUSR1);
+        }
+        else
+            counter++;
     } else if (sig == SIGUSR1) {
         if (counter == 0) {
             my_putstr("successfully connected\n\n");
